Finish first draft of landing page project for CSS
Helpful approach was first taking time to conceptually lay out the
various elements on the page, as well as the containers that are needed
to make it work. Sometimes, it can seem like we are adding too many
containers, usually in the form of <divs>, but they typically provide
greater control of spacing between and within the various flex items.
Another tip to self is to consider using more % based spacing
measurements, such as for padding or margin, so object can move and
scale.

Need to figure out how to streamline CSS arguments. Currently writing
new rules for every single flex item, and the flow is chaotic. It seems
like the flow should be grouped together, rather than tracked to the
order that elemenets appear in the html.

Also included screenshot of the completed product - may make adjustments
or use as base reference to build off and tinker with.